src/main/java/com/example/books/service/MyUserDetailsService.java
package com.example.books.service;

import com.example.books.model.MyUserDetails;
import com.example.books.model.User;
import com.example.books.repository.UserRepository;
import com.example.books.util.ErrorMessages;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.context.annotation.Lazy;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.Optional;

/**
 * Service handling the user related operations.
 */
@Service
public class MyUserDetailsService implements UserDetailsService {

    private final UserRepository userRepository;
    private final PasswordEncoder passwordEncoder;

    @Autowired
    public MyUserDetailsService(UserRepository userRepository, @Lazy PasswordEncoder passwordEncoder) {
        this.userRepository = userRepository;
        this.passwordEncoder = passwordEncoder;
    }

    /**
     * Creates an MyUserDetails object based on the User found on the repository using username
     * @param username String
     * @return UserDetails object
     * @throws UsernameNotFoundException on errors
     */
    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        Optional<User> user = userRepository.findByUsername(username);
        user.orElseThrow(() -> new UsernameNotFoundException(ErrorMessages.NOT_FOUND + username));
        return user.map(MyUserDetails::new).get();
    }

    /**
     * Creates a user.
     * @param user User
     * @return User
     */
    public User save(User user) {
        user.setPassword(passwordEncoder.encode(user.getPassword()));
        return userRepository.save(user);
    }

    /**
     * gets current logged user.
     * @return String
     */
    public String getCurrentUsername() {
        return SecurityContextHolder.getContext().getAuthentication().getName();
    }
}